+++ ci/install_osx.sh
@@ -1,2 +1,20 @@
+brew tap dartsim/dart # for ipopt
+
 brew update > /dev/null
-brew bundle
+
+brew install git
+brew install cmake
+brew install assimp
+brew install fcl
+brew install bullet --with-double-precision
+brew install ode --with-libccd --with-double-precision
+brew install flann
+brew install boost
+brew install eigen
+brew install tinyxml
+brew install tinyxml2
+brew install libccd
+brew install nlopt
+brew install ipopt
+brew install urdfdom
+brew install open-scene-graph