+++ dart/dynamics/detail/JointPtr.hpp
@@ -127,7 +127,7 @@ public:
   /// Equality
   template <class OtherJointT, class OtherBodyNodeT>
   bool operator == (const TemplateJointPtr<OtherJointT,
-                    OtherBodyNodeT>& _rhs)
+                    OtherBodyNodeT>& _rhs) const
   {
     return mBodyNodePtr == _rhs.mBodyNodePtr;
   }
@@ -135,7 +135,7 @@ public:
   /// Inequality
   template <class OtherJointT, class OtherBodyNodeT>
   bool operator != (const TemplateJointPtr<OtherJointT,
-                    OtherBodyNodeT>& _rhs)
+                    OtherBodyNodeT>& _rhs) const
   {
     return !( *this == _rhs );
   }
@@ -143,7 +143,7 @@ public:
   /// Less than
   template <class OtherJointT, class OtherBodyNodeT>
   bool operator < (const TemplateJointPtr<OtherJointT,
-                   OtherBodyNodeT>& _rhs)
+                   OtherBodyNodeT>& _rhs) const
   {
     return (mBodyNodePtr < _rhs.mBodyNodePtr);
   }
@@ -151,7 +151,7 @@ public:
   /// Greater than
   template <class OtherJointT, class OtherBodyNodeT>
   bool operator > (const TemplateJointPtr<OtherJointT,
-                   OtherBodyNodeT>& _rhs)
+                   OtherBodyNodeT>& _rhs) const
   {
     return (mBodyNodePtr > _rhs.mBodyNodePtr);
   }
@@ -159,7 +159,7 @@ public:
   /// Less than or equal to
   template <class OtherJointT, class OtherBodyNodeT>
   bool operator <= (const TemplateJointPtr<OtherJointT,
-                    OtherBodyNodeT>& _rhs)
+                    OtherBodyNodeT>& _rhs) const
   {
     return (*this < _rhs) || (*this == _rhs);
   }
@@ -167,7 +167,7 @@ public:
   /// Greater than or equal to
   template <class OtherJointT, class OtherBodyNodeT>
   bool operator >= (const TemplateJointPtr<OtherJointT,
-                    OtherBodyNodeT>& _rhs)
+                    OtherBodyNodeT>& _rhs) const
   {
     return (*this > _rhs) || (*this == _rhs);
   }