+++ dart/gui/osg/Viewer.cpp
@@ -380,6 +380,9 @@ void Viewer::addWorldNode(WorldNode* _newWorldNode, bool _active)
     _newWorldNode->simulate(mSimulating);
   _newWorldNode->mViewer = this;
   _newWorldNode->setupViewer();
+  // set again the shadow technique to produce warning for ImGuiViewer
+  if(_newWorldNode->isShadowed())
+    _newWorldNode->setShadowTechnique(_newWorldNode->getShadowTechnique());
 }
 
 //==============================================================================
@@ -466,6 +469,15 @@ const std::unordered_set<ViewerAttachment*>& Viewer::getAttachments() const
 const ::osg::Group* Viewer::getLightGroup() const
 {
   return mLightGroup;
+}
+
+//==============================================================================
+const ::osg::ref_ptr<::osg::LightSource>& Viewer::getLightSource(std::size_t index) const
+{
+  assert(index < 2);
+  if(index == 0)
+    return mLightSource1;
+  return mLightSource2;
 }
 
 //==============================================================================