+++ dart/gui/osg/WorldNode.hpp
@@ -34,6 +34,7 @@
 #define DART_GUI_OSG_WORLDNODE_HPP_
 
 #include <osg/Group>
+#include <osgShadow/ShadowTechnique>
 #include <unordered_map>
 #include <memory>
 
@@ -67,10 +68,11 @@ public:
   friend class Viewer;
 
   /// Default constructor
-  explicit WorldNode(std::shared_ptr<dart::simulation::World> _world = nullptr);
+  /// Shadows are disabled by default
+  explicit WorldNode(std::shared_ptr<dart::simulation::World> world = nullptr, ::osg::ref_ptr<osgShadow::ShadowTechnique> shadowTechnique = nullptr);
 
   /// Set the World that this WorldNode is associated with
-  void setWorld(std::shared_ptr<dart::simulation::World> _newWorld);
+  void setWorld(std::shared_ptr<dart::simulation::World> newWorld);
 
   /// Get the World that this WorldNode is associated with
   std::shared_ptr<dart::simulation::World> getWorld() const;
@@ -119,15 +121,30 @@ public:
 
   /// Pass in true to take steps between render cycles; pass in false to turn
   /// off steps between render cycles.
-  void simulate(bool _on);
+  void simulate(bool on);
 
   /// Set the number of steps to take between each render cycle (only if the
   /// simulation is not paused)
-  void setNumStepsPerCycle(std::size_t _steps);
+  void setNumStepsPerCycle(std::size_t steps);
 
   /// Get the number of steps that will be taken between each render cycle (only
   /// if the simulation is not paused)
   std::size_t getNumStepsPerCycle() const;
+
+  /// Get whether the WorldNode is casting shadows
+  bool isShadowed() const;
+
+  /// Set the ShadowTechnique
+  /// If you wish to disable shadows, pass a nullptr
+  void setShadowTechnique(::osg::ref_ptr<osgShadow::ShadowTechnique> shadowTechnique = nullptr);
+
+  /// Get the current ShadowTechnique
+  /// nullptr is there are no shadows
+  ::osg::ref_ptr<osgShadow::ShadowTechnique> getShadowTechnique() const;
+
+  /// Helper function to create a default ShadowTechnique given a Viewer
+  /// the default ShadowTechnique is ShadowMap
+  static ::osg::ref_ptr<osgShadow::ShadowTechnique> createDefaultShadowTechnique(const Viewer* viewer);
 
 protected:
 
@@ -172,6 +189,15 @@ protected:
 
   /// Viewer that this WorldNode is inside of
   Viewer* mViewer;
+
+  /// OSG group for non-shadowed objects
+  ::osg::ref_ptr<::osg::Group> mNormalGroup;
+
+  /// OSG group for shadowed objects
+  ::osg::ref_ptr<::osg::Group> mShadowedGroup;
+
+  /// Whether the shadows are enabled
+  bool mShadowed;
 
 };